[LED-113] Removing Duplicate Code
