+++ src/main/java/ledger/io/input/USBankQFXConverter.java
@@ -21,93 +21,13 @@ import java.util.regex.Pattern;
 /**
  * Handles the converting of Quicken Qxf files of the old format into our internal transaction objects.
  */
-public class USBankQFXConverter implements IInAdapter<Transaction> {
-
-    private File qfxFile;
-    private Account account;
+public class USBankQFXConverter extends AbstractUFXConverter {
 
     public USBankQFXConverter(File file, Account account) {
-        this.qfxFile = file;
-        this.account = account;
+        super(file,account);
     }
 
-    /**
-     * Parses the given file into the application's internal transaction objects.
-     *
-     * @return A list of Transaction objects created from the transactions in the provided file.
-     * @throws IOException When unable to read the given file
-     */
-    @Override
-    public List<Transaction> convert() throws ConverterException {
-        List<Transaction> transactions = new ArrayList();
-
-        // read in given file
-        String sgml = null;
-        try {
-            sgml = new Scanner(qfxFile).useDelimiter("\\Z").next();
-        } catch (FileNotFoundException e) {
-            throw new ConverterException("The QFX file could not be found.", e);
-        }
-
-        // chop off everything before and after transactions (before/after STMTTRN)
-        int indexOfFirstTrans = sgml.indexOf("<STMTTRN>");
-        sgml = sgml.substring(indexOfFirstTrans);
-
-        int lastIndexOfTrans = sgml.indexOf("</BANKTRANLIST>");
-        if(lastIndexOfTrans == -1) throw new ConverterException("The provided QFX file is malformed.", new IndexOutOfBoundsException());
-        sgml = sgml.substring(0, lastIndexOfTrans);
-
-        StringBuilder correctedXml = correctXml(sgml);
-
-        DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-        DocumentBuilder builder;
-        try {
-            builder = factory.newDocumentBuilder();
-        } catch (ParserConfigurationException e) {
-            throw new ConverterException("Unable to create new XML parser.", e);
-        }
-        InputSource is = new InputSource(new StringReader(correctedXml.toString()));
-        Document xml;
-        try {
-            xml = builder.parse(is);
-        } catch (SAXException e) {
-            throw new ConverterException("Unable to parse the given file.", e);
-        } catch (IOException e) {
-            throw new ConverterException("Unable to parse the given file.", e);
-        }
-        parseXml(transactions, xml);
-
-        // return list
-        return transactions;
-    }
-
-    private StringBuilder correctXml(String sgml) {
-        // add tags to the all rows that are not STMTTRN
-        String[] splitPieces = sgml.split("<");
-        LinkedList<String> modifiedPieces = new LinkedList();
-        for (String piece : splitPieces) {
-            piece = piece.trim();
-            boolean matches = Pattern.matches("(?!STMTTRN[>].*|[/]STMTTRN[>].*).*[>].*", piece);
-            if (matches) {
-                int lastClosingTag = piece.indexOf(">");
-                String xmlTag = piece.substring(0, lastClosingTag);
-                String correctXml = "<" + piece + "</" + xmlTag + ">";
-                modifiedPieces.add(correctXml + "\n");
-            } else if (!piece.equals("")) {
-                modifiedPieces.add("<" + piece + "\n");
-            }
-        }
-
-        StringBuilder correctedXml = new StringBuilder();
-        correctedXml.append("<?xml version=\"1.0\" encoding=\"UTF-8\"?>" + "\n" + "<QFX>");
-        for (String piece : modifiedPieces) {
-            correctedXml.append(piece);
-        }
-        correctedXml.append("\n" + "</QFX>");
-        return correctedXml;
-    }
-
-    private void parseXml(List<Transaction> transactions, Document xml) throws ConverterException {
+    protected void parseXml(List<Transaction> transactions, Document xml) throws ConverterException {
         //parse xml
         NodeList transactionTypes = xml.getElementsByTagName("TRNTYPE");
         NodeList transactionDates = xml.getElementsByTagName("DTPOSTED");
@@ -143,7 +63,7 @@ public class USBankQFXConverter implements IInAdapter<Transaction> {
                 List<Tag> tags = new LinkedList<Tag>();
                 Note note = new Note(memos.item(i).getTextContent());
 
-                Transaction transaction = new Transaction(date, type, amount, this.account, payee, false, tags, note);
+                Transaction transaction = new Transaction(date, type, amount, this.getAccount(), payee, false, tags, note);
                 transactions.add(transaction);
             }
         } catch (NullPointerException e) {
